+++ lib/core/routing/app_router.dart
@@ -8,7 +8,7 @@ import 'package:shopping_app/features/sign_up/ui/sign_up_screen.dart';
 class AppRouter {
   Route? generateRoute(RouteSettings settings) {
     //this arguments to be passed in any screen like this ( arguments as ClassName )
-    final arguments = settings.arguments;
+   // final arguments = settings.arguments;
 
     switch (settings.name) {
       case Routes.loginScreen:
