+++ lib/features/sign_up/ui/widgets/sign_up_form.dart
@@ -1,4 +1,3 @@
-import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
